move throwing variant to test only place
